Add View::style(), #[derive(IntoViewElement)], and EditorView
