+++ crates/gpui/examples/text_views/main.rs
@@ -28,7 +28,6 @@
 //! ```
 
 mod editor;
-mod editor_text;
 mod input;
 mod text_area;
