+++ crates/gpui/src/gpui.rs
@@ -90,7 +90,8 @@ pub use executor::*;
 pub use geometry::*;
 pub use global::*;
 pub use gpui_macros::{
-    AppContext, IntoElement, Render, VisualContext, property_test, register_action, test,
+    AppContext, IntoElement, IntoViewElement, Render, VisualContext, property_test,
+    register_action, test,
 };
 pub use gpui_util::arc_cow::ArcCow;
 pub use http_client;
