+++ crates/gpui_macros/src/gpui_macros.rs
@@ -1,6 +1,7 @@
 mod derive_action;
 mod derive_app_context;
 mod derive_into_element;
+mod derive_into_view_element;
 mod derive_render;
 mod derive_visual_context;
 mod property_test;
@@ -33,6 +34,14 @@ pub fn register_action(ident: TokenStream) -> TokenStream {
 #[proc_macro_derive(IntoElement)]
 pub fn derive_into_element(input: TokenStream) -> TokenStream {
     derive_into_element::derive_into_element(input)
+}
+
+/// #[derive(IntoViewElement)] generates an `IntoElement` implementation for types
+/// that implement the `View` trait. It creates a `ViewElement` wrapper and
+/// automatically enables caching based on the `View::style()` method.
+#[proc_macro_derive(IntoViewElement)]
+pub fn derive_into_view_element(input: TokenStream) -> TokenStream {
+    derive_into_view_element::derive_into_view_element(input)
 }
 
 #[proc_macro_derive(Render)]
